Initial version of exception and error handling.

+++ src/main/java/bg/beesoft/beehive/service/ApiaryService.java
@@ -6,10 +6,12 @@ import bg.beesoft.beehive.model.entity.AddressEntity;
 import bg.beesoft.beehive.model.entity.ApiaryEntity;
 import bg.beesoft.beehive.model.entity.BeehiveEntity;
 import bg.beesoft.beehive.model.entity.UserEntity;
+import bg.beesoft.beehive.model.exception.UnauthorizedRequestException;
 import bg.beesoft.beehive.model.view.ApiaryView;
 import bg.beesoft.beehive.model.view.BeehiveView;
 import bg.beesoft.beehive.repository.ApiaryRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
@@ -33,6 +35,7 @@ public class ApiaryService {
     }
 
     public void save(ApiaryAddDTO apiaryAddDTO, UserDetails userDetails) {
+
         ApiaryEntity apiaryEntity = new ApiaryEntity().
                 setName(apiaryAddDTO.getName())
                 .setDescription(apiaryAddDTO.getDescription())
@@ -63,19 +66,29 @@ public class ApiaryService {
                 .map(e -> modelMapper.map(e, ApiaryView.class))
                 .collect(Collectors.toList());
 
-//        apiaries.stream().forEach(ap-> apiaryViews.get(apiaries.indexOf(ap))
-//                .setBeehives(ap.getBeehives().stream()
-//                        .map(b->modelMapper.map(b, BeehiveView.class)).collect(Collectors.toList())));
 
     return apiaryViews;
     }
 
-    public ApiaryEditDTO findById(Long id) {
-        return modelMapper.map(apiaryRepository.findById(id).orElse(null), ApiaryEditDTO.class);
+    public ApiaryEditDTO findById(Long id,UserDetails userDetails) {
+        ApiaryEntity apiaryEntity = apiaryRepository.findById(id).orElseThrow();
+
+        if (!apiaryEntity.getBeekeeper().getEmail().equals(userDetails.getUsername())){
+            throw new UnauthorizedRequestException("Нямате достъп до този пчелин.");
+        }
+
+        return  modelMapper.map(apiaryEntity, ApiaryEditDTO.class);
     }
 
-    public void update(ApiaryEditDTO apiaryEditDTO, Optional<AddressEntity> optionalAddress) {
-        ApiaryEntity apiaryEntity = apiaryRepository.findById(apiaryEditDTO.getId()).orElse(new ApiaryEntity()).
+    public void update(ApiaryEditDTO apiaryEditDTO, Optional<AddressEntity> optionalAddress,UserDetails userDetails) {
+        ApiaryEntity apiaryEntity = apiaryRepository.findById(apiaryEditDTO.getId()).orElseThrow();
+
+        if (!apiaryEntity.getBeekeeper().getEmail().equals(userDetails.getUsername())){
+            throw new UnauthorizedRequestException("Нямате достъп до този пчелин.");
+        }
+
+
+        apiaryEntity.
                 setName(apiaryEditDTO.getName())
                 .setDescription(apiaryEditDTO.getDescription())
                 .setArea(apiaryEditDTO.getArea())
@@ -100,7 +113,10 @@ public class ApiaryService {
         return apiaryRepository.findById(id).get().getAddress().getId();
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, UserDetails userDetails) {
+        if (!apiaryRepository.findById(id).orElseThrow().getBeekeeper().getEmail().equals(userDetails.getUsername())){
+            throw new UnauthorizedRequestException("Нямате достъп до този пчелин.");
+        }
         apiaryRepository.deleteById(id);
     }
 
@@ -129,8 +145,8 @@ public class ApiaryService {
         apiaryRepository.save(apiaryEntity);
     }
 
-    public ApiaryView findViewById(Long id) {
-        return modelMapper.map(findById(id), ApiaryView.class);
+    public ApiaryView findViewById(Long id, UserDetails userDetails) {
+        return modelMapper.map(findById(id,userDetails), ApiaryView.class);
     }
 
     public Long findIdByReferenceNumberInApiary(Long apiaryId, int referenceNumber) {
@@ -139,5 +155,9 @@ public class ApiaryService {
                 getBeehives().stream().
                 filter(b -> b.getReferenceNumber() == referenceNumber)
                 .map(b -> b.getId()).findFirst().orElse(null);
+    }
+
+    public ApiaryEntity findById(Long apiaryId) {
+    return apiaryRepository.findById(apiaryId).orElseThrow();
     }
 }
